refactor: moved ProductsFilter
